get user all, admins, drinkers

+++ app/Services/CountryService.php
@@ -5,43 +5,61 @@ namespace App\Services;
 
 
 use App\Models\Country;
+use App\Repositories\GetCountriesRepository;
+use App\Repositories\StoreCountryRepository;
 use App\Services\Interfaces\CountryServiceInterface;
 use App\Traits\ResponseDataTrait;
 
 class CountryService implements CountryServiceInterface
 {
     use ResponseDataTrait;
+    private $getcountriesRepository;
+    private $storecountryRepository;
+
+    public function __construct(GetCountriesRepository $getcountriesRepository,
+    StoreCountryRepository $storecountryRepository)
+    {
+        $this->getcountriesRepository = $getcountriesRepository;
+        $this->storecountryRepository = $storecountryRepository;
+    }
 
     public function getAll()
     {
-        $countries = Country::with('beers')->get();
+        $countries = $this->getcountriesRepository->getAll();
 
         return $this->responseWithData($countries, 200);
     }
 
     public function getCountry($country)
     {
-        $countries = Country::with('beers')
-            ->where('id', $country->id)
-            ->get();
-
+        $countries = $country
+            ? $this->getcountriesRepository->getCountry($country)
+            : false;
         return $this->responseWithData($countries, 200);
     }
 
     public function storeNew($data)
     {
-        $country = new Country();
-
-        $country->name = $data->name;
-        $country->shortcut = $data->shortcut;
-
-        $country->save();
-        return $this->responseWithData($country, 200);
+        if (auth()->user()->hasRole('admin'))
+        {
+            $country = $data
+                ? $this->storecountryRepository->addCountry($data)
+                : false;
+            return $this->responseWithData($country, 200);
+        } else {
+            return $this->responseWithMessage('Unauthorized', 401);
+        }
     }
 
     public function remove($country)
     {
-        $country->delete();
-        return $this->responseWithMessage('Country has been removed', 200);
+        if (auth()->user()->hasRole('admin'))
+        {
+            return $country->delete()
+                ? $this->responseWithMessage('Country has been removed', 200)
+                : $this->responseWithMessage('Not found', 404);
+        } else {
+            return $this->responseWithMessage('Unauthorized', 401);
+        }
     }
 }
